#126 rearrange tests for create vendor from anitbody creation
